+++ MisticaCatalog/Source/Catalog/UICatalogViewController.swift
@@ -3,7 +3,7 @@
 //
 //  Made with ❤️ by Novum
 //
-//  Copyright © 2020 Telefonica. All rights reserved.
+//  Copyright © Telefonica. All rights reserved.
 //
 
 import Mistica
@@ -15,6 +15,7 @@ private enum UICatalogRow: Int, CaseIterable {
     case checkbox
     case controls
     case crouton
+    case dataCard
     case feedbacks
     case fonts
     case forms
@@ -22,6 +23,7 @@ private enum UICatalogRow: Int, CaseIterable {
     case highlightedCard
     case inputFields
     case lists
+    case mediaCard
     case radioButton
     case scrollContentIndicator
     case sectionTitle
@@ -121,7 +123,7 @@ extension UICatalogViewController: UITableViewDataSource, UITableViewDelegate {
     }
 
     public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: Constants.cellReusableIdentifier, for: indexPath) as! ListCellView
+        let cell = tableView.dequeueReusableCell(withIdentifier: Constants.cellReusableIdentifier, for: indexPath) as! ListViewCell
 
         let catalogRow = UICatalogRow(rawValue: indexPath.row)!
 
@@ -175,6 +177,10 @@ extension UICatalogViewController: UITableViewDataSource, UITableViewDelegate {
             show(UICatalogHighlightedCardViewController(), sender: self)
         case .checkbox:
             show(UICatalogCheckboxViewController(), sender: self)
+        case .mediaCard:
+            show(UICatalogMediaCardViewController(), sender: self)
+        case .dataCard:
+            show(UICatalogDataCardViewController(), sender: self)
         case .stepper:
             show(UICatalogStepperViewController(), sender: self)
         }
@@ -197,7 +203,7 @@ private extension UICatalogViewController {
 
         tableView.dataSource = self
         tableView.delegate = self
-        tableView.register(ListCellView.self, forCellReuseIdentifier: Constants.cellReusableIdentifier)
+        tableView.register(ListViewCell.self, forCellReuseIdentifier: Constants.cellReusableIdentifier)
         tableView.sectionHeaderHeight = UITableView.automaticDimension
         tableView.estimatedSectionHeaderHeight = 80
 
@@ -271,6 +277,10 @@ private extension UICatalogRow {
             return "Controls"
         case .highlightedCard:
             return "HighlightedCard"
+        case .mediaCard:
+            return "MediaCard"
+        case .dataCard:
+            return "DataCard"
         case .stepper:
             return "Stepper"
         }
@@ -314,8 +324,8 @@ private extension UICatalogRow {
             return .headerIcon
         case .controls:
             return .controlsIcon
-        case .highlightedCard:
-            return .highlightedCardIcon
+        case .highlightedCard, .mediaCard, .dataCard:
+            return .cardIcon
         case .stepper:
             return .stepperIcon
         }