+++ MisticaCatalog/Source/ComponentsCatalogs/UICatalogStepperViewController.swift
@@ -3,7 +3,7 @@
 //
 //  Made with ❤️ by Novum
 //
-//  Copyright © 2020 Telefonica. All rights reserved.
+//  Copyright © Telefonica. All rights reserved.
 //
 
 import Foundation
@@ -19,7 +19,7 @@ private enum Section: Int, CaseIterable {
 class UICatalogStepperViewController: UITableViewController {
     private var numberOfSteps = 3
     private var minimumNumberOfSteps = 2
-    private var maximumNumberOfSteps = 6
+    private var maximumNumberOfSteps = 5
     private var currentStep = 0
 
     private var indeterminateStepperView: IndeterminateStepperView!
@@ -32,6 +32,10 @@ class UICatalogStepperViewController: UITableViewController {
     private let numberOfStepsStepper = UIStepper()
     private let numberOfStepsLabel = UILabel()
     private lazy var numberOfStepsStackView = UIStackView(arrangedSubviews: [numberOfStepsLabel, numberOfStepsStepper])
+
+    private let animatedSwitch = UISwitch()
+    private let animatedLabel = UILabel()
+    private lazy var animatedStackView = UIStackView(arrangedSubviews: [animatedLabel, animatedSwitch])
 
     init() {
         if #available(iOS 13.0, *) {
@@ -48,7 +52,7 @@ class UICatalogStepperViewController: UITableViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        title = "Checkbox"
+        title = "Stepper"
 
         tableView.dataSource = self
         tableView.delegate = self
@@ -101,8 +105,15 @@ extension UICatalogStepperViewController {
             numberOfStepsStackView.distribution = .equalSpacing
             numberOfStepsStackView.alignment = .center
 
+            // Animated
+            animatedLabel.text = "Animated"
+            animatedSwitch.isOn = true
+
+            animatedStackView.distribution = .equalSpacing
+            animatedStackView.alignment = .center
+
             // Vertical stack view
-            let stackView = UIStackView(arrangedSubviews: [numberOfStepsStackView, currentStepStackView])
+            let stackView = UIStackView(arrangedSubviews: [numberOfStepsStackView, currentStepStackView, animatedStackView])
             stackView.axis = .vertical
             stackView.spacing = 8
 
@@ -137,8 +148,8 @@ extension UICatalogStepperViewController {
     }
 
     func updateStepperValues() {
-        indeterminateStepperView.setValue(Int(Float(currentStep) / Float(numberOfSteps) * 100), animated: true)
-        determinateStepperView.setCurrentStep(currentStep, animated: true)
+        indeterminateStepperView.setValue(Int(Float(currentStep) / Float(numberOfSteps) * 100), animated: animatedSwitch.isOn)
+        determinateStepperView.setCurrentStep(currentStep, animated: animatedSwitch.isOn)
     }
 }